+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+This release introduces the following breaking changes:
+
+- feat(sumologicextension): use hostname as default collector name [#918]
+
+[#918]: https://github.com/SumoLogic/sumologic-otel-collector/pull/918
+[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.70.0-sumo-0...main
+
 ## [v0.70.0-sumo-0]
 
 ### Added